Update js/jquery.coda-slider-3.0.js
I mostly want to add the arrows on the slider panels and not want them near the sliders (with floats or so). With this change the positioning of the coda-nav-left and right is checked. When you positionate them absolute they will not need to increase the width of the slider-wrapper.